+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ## [Unreleased]
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- Detect cortex.yaml and show a one-time onboarding tip [MVP-1]